+++ astroquery/mast/observations.py
@@ -11,9 +11,11 @@ import warnings
 import time
 import os
 from urllib.parse import quote
+import importlib.util
 
 import numpy as np
 import astropy.units as u
+from astropy.io import fits
 import astropy.coordinates as coord
 from requests import HTTPError
 from astropy.table import Table, Row, vstack
@@ -35,6 +37,13 @@ try:
     from botocore.exceptions import ClientError, BotoCoreError
 except ImportError:
     pass
+
+try:
+    import asdf
+    import fsspec
+except ImportError:
+    asdf = None
+    fsspec = None
 
 __all__ = ['Observations', 'ObservationsClass', 'MastClass', 'Mast']
 
@@ -1202,6 +1211,75 @@ class ObservationsClass(MastQueryWithLogin):
         if len(unique_products) < len(products):
             log.info("To return all products, use `Observations.get_product_list`")
         return unique_products
+
+    # TODO: Need to inlcude way to parse if it is a MAST on prem URL and handle the streaming of that
+    def read_product(self, product_path, ignore_unrecognized=True, **kwargs):
+        """
+        Read a product from Open S3 bucket to memory. Currently supports FITS and ASDF product types only.
+
+        Parameters
+        ----------
+        product_path: str
+            URI to the product in the STScI S3 open data bucket.
+        ignore_unrecognized: bool
+            Tells asdf.open() to include or ignore warnings from unrecognized asdf tags. Defaults to True
+        **kwargs
+            Additional keyword arguments passed to the underlying file reader:
+            - For FITS files: forwarded to ``astropy.io.fits.open``.
+            Common options include ``memmap``, ``mode``, etc.
+            - Ignored for ASDF files (except for future extension if needed).
+
+        Returns
+        -------
+        object
+            FITS or ASDF object for the given data product.
+        """
+        # Checks if a path is empty or None.
+        if not product_path or not str(product_path).strip():
+            raise ValueError("No product path provided")
+
+        # Forces the path to be lowercase for the extension checks. This is only used for the checks
+        path = str(product_path).lower()
+
+        # Checks users enviornment for fsspec, required for both fits and asdf
+        if fsspec is None:
+            raise ImportError('The "fsspec" package is required to read products directly from a URI. '
+                              'Please install it with `pip install fsspec`.')
+
+        # Logic for reading FITS files
+        if path.endswith((".fits", ".fits.gz")):
+            try:
+                data_product = fits.open(product_path, fsspec_kwargs={"anon": True}, **kwargs)
+                log.info(f"Loaded: {product_path}")
+                return data_product
+            except Exception as e:
+                raise RuntimeError(f"Failed to open FITS File: {product_path} {e}")
+
+        # Logic for reading ASDF files
+        elif path.endswith(".asdf"):
+            # checks for asdf package and will raise and error if not installed as asdf is required
+            for pkg in ["asdf"]:
+                if importlib.util.find_spec(pkg) is None:
+                    raise ImportError(f'The "{pkg}" package is required to read ASDF files containing {pkg} data. '
+                                      f'Please install it with `pip install {pkg}`.')
+
+            # Checks for gwcs and warns the user if it is not installed, this will not stop the function.
+            for pkg in ["gwcs"]:
+                if importlib.util.find_spec(pkg) is None:
+                    warnings.warn(f'The "{pkg}" package is required to read ASDF files containing {pkg} data. '
+                                  f'Please install it with `pip install {pkg}`.')
+
+            # Attempts to open the asdf files
+            try:
+                f = fsspec.open(product_path, "rb", anon=True).open()
+                data_product = asdf.open(f, ignore_unrecognized_tag=ignore_unrecognized)
+                log.info(f"Loaded: {product_path}")
+                return data_product
+            except Exception as e:
+                raise RuntimeError(f"Failed to open ASDF File: {product_path} {e}")
+
+        else:
+            raise ValueError(f"Unsupported product type: {product_path}")
 
 
 @async_to_sync